Fix line diff analysis on lines with non-ASCII characters
Previously, the UTF-8 encoded byte sequence was diffed.  This works well
enough on ASCII input, and on some non-ASCII input, but risks splitting
up multibyte characters when inserting the diff markup later on, if the
differ decides that some but not all bytes of a multibyte character were
modified.

To avoid this, decode the lines as UTF-8 first so that we get unicode
strings to do the diffing on.  Produce byte offsets as the final result
by encoding the unicode strings as UTF-8 again at the end.