Improve UI when commenting old version of modified lines
There's a warning when commenting lines that have been modified by later
commits in the review.  The reason is that an issue raised against such
lines will be open, but it won't be present in the head commit, and can
never be marked as addressed automatically, thus requiring manual
resolution at some point.

This warning was also triggered when commenting the old version of lines
modified in the current diff / by the current commit.  While technically
not inaccurate, the warning message was unhelpful.  Instead of that, just
display an inline warning about the consequences of commenting modified
lines.

Also improve the warning shown when commenting lines actually modified by
a later commit by including a brief explanation of what the problem with
such comments is.

Finally, a drive-by fix: don't start "comment marking" if we're already
doing it.  Doing so causes zombie comment markers to remain.  This happens
when multiple sets of event handlers are registered.  Normally, that
shouldn't happen, but there are cases when it does, and it's easy enough
to make such redundant event handlers harmless.