Use only absolute paths in client-side scripts
In some places absolute paths were already used, but far from everywhere.
Using relative paths creates an unfortunate dependency on the path of the
page, which is often varying (review front-page could be either /r/NNN or
/showreview?id=NNN, which makes a huge difference for relative paths.)

This hasn't been a problem because (almost) all pages are delivered with a

  <base href="/">

element, but that creates its own problems and is quite ugly.