Use absolute paths on /showcommit page; drop <base href="/"> element
Having the <base href="/"> element causes the path name links in the list
of modified files to be to /#fNNN instead of to the named anchors on the
page.  This also means that clicking such a link before the script that
sets event handlers on the links has run navigates to / by mistake.