Discontinue 'knownhosts' table
It was used to track hosts that the Critic system had passwordless SSH
access to, and thus could access repositories from.  But there was no
UI for adding rows to the table, leaving some functionality (adding
tracked branches on /repositories, and creating tracking reviews via
/createreview) effectively disabled.

Keeping track of repository URLs that are valid is rather pointless in
practice.  We can just let the user enter a repository URL, and if it
works, it works.