Don't abort install/upgrade if restarting Apache fails
At least not unless the users asks us to.  If one of the modules
that provide SSL support (mod_ssl/mod_gnutls) is enabled, and
Critic is installed with one of the access schemes "https" or
"both", Apache will typically fail to restart due to the SSL
configuration being incomplete.  Unconditionally aborting the
install then is not very useful, since all the user needs to do
is to complete the configuration.