Treat each space character as its own white-space "word"
When analyzing lines with only white-space changes, groups of up to four
space characters were treated as "words."  I don't know why, really.  The
effect of this was that if a space-indented line had its indentation
changed by an amount that wasn't a multiple of four in certain ways, it
would be represented in the inter-line diff as if some space characters
in the old version were replaced by other space characters in the new
version.

For instance, if a line had its indentation changed from 8 spaces to
6 spaces, the inter-line diff would be that characters 4-7 on the old line
were replaced by characters 4-5 on the new line, and that characters 0-3
where unchanged.

By treating each space character as a "word," the inter-line diff will
instead simply be that characters 0-1 were deleted, which looks much
nicer.  It looks much nicer in particular when a block of lines are re-
indented by the same amount, since this is then represented as a ribbon
of red or green at column zero on one of the sides of the diff, while the
other side is "clean."

For (purely) tab-indented code this is how it already worked, since tab
characters were already treated as individual words.