Increment review serial number when rebasing
For a move rebase it was typically done anyway as part of adding
a commit (equivalent merge or replayed rebase) to the review, even
when that commit is "empty" and not later displayed in the list of
commits on the review front-page.  (Incrementing twice is not a
problem.)

For a history rewrite rebase, however, it was not typically done.