Show the error, not HTTP 500 when commit range starts with merge commit
WrappedResult doesn't do "value = str(value)" in next() so the
generators currently have to emit strings. It might be useful
with generators that emit non-strings for something in the
future but a generator of htmlutils.Document objects makes little
sense so convert to more modern DisplayMessage with a capital D
instead.

Repro steps for the HTTP 500 were:
1. add critic as a repo in critic
2. /showcommit?first=57a886e6352b229991c81e7ba43244ace7e02d76&last=b2b78ca013b49c73231bee11674bcdb3edf6d3f2