Define REMOTE_USER when pushing, and stop guessing
When the Critic system user updates a branch, the git hook would
try to guess which user was actually pushing by looking at the
commit meta data (specifically the committer email address) of the
new head of the updated branch.  This is obviously flawed in all
sorts of ways, and would cause crashes in some cases if no user
could be identified that way.

Instead of doing this, define REMOTE_USER when pushing (both via
the /rebasereview operation and via extensions) as the Critic
system user, and stop guessing in the git hook.

Also adjust the extension repository work copy mechanism to be
unavailable if the user is anonymous.  Anonymous users can only use
extensions that the system's administrator has installed
universally (which ought to mean the system administrator trusts
the extension, but since work copies are kept semi-persistently per
user, it sort of doesn't make sense to allow them for anonymous
users.