Offer both [download] and [view] links on /showfile
Earlier the approach was to try to guess if the user wanted to view a
file in the browser or if they wanted to download the file, based on
the file extension. For this to work perfectly the list in
configuration/mimetypes.py has to include almost all file types there
is and this list if both very long and a moving target. Further, some
extensions might be used for two different things, one which is
essentially plain text that can be viewed in a browser and one which
is some kind of binary format. So instead of trying to figure it out,
just include two links so that the user can choose to download or
view each file. We still allow for explicit mime types to be
configured in configuration/mimetypes.py though because we still want
them for the various different kinds of types that the browser can
view like image/png etc.