Fix minor issue when generating text diff for added/removed file
The per-file diff starts with the lines

  --- a/<path>
  +++ b/<path>

where /dev/null is used as the path on one of the lines if the file was
added or removed.  The conditions for using /dev/null were reversed, so
if the file was removed, the path on the --- line was /dev/null, and if
the file was added, the path on the +++ line was /dev/null.