Output test summary after running tests
Makes it possible to quickly see if any tests failed or any errors
or warnings were logged.  In particular when running with --debug,
such information is easily drowned among all the other output.

The majority of the changes are related to wrapping the logger
in a wrapper class that counts errors and warnings.  It also has
the bonus effect that it converts (the effect of)

  logger.debug("foo\nbar")

into (that of)

  logger.debug("foo")
  logger.debug("bar")

which produces a more uniform output.