Don't do syntax highlighting when adding commits to review
It'll be done on demand later on when someone loads the diff anyway.
If many large commits are added to a review at once, and then never
displayed individually (i.e. only squashed together) much of that
syntax highlighting may have been wasted.  Also, doing it all when
the commits are added creates a long wait for the user pushing the
commits.

And finally, it slows down testing (e.g. 012-replayrebase.py) for
no good reason.