Avoid line-breaks in OperaCritic-URL header value in emails
If a user has requested that multiple URL types be used in emails (using
the email.urlType preference) they are normally separated by line-breaks
and indented.  In the OperaCritic-URL header value they were just separated
by line-breaks, and not indented.  This is not a possible header value, so
either the header value will be considered invalid, or it'll be modified,
for instance by injecting a space character after each line-break.

To avoid this impropriety, separate the different URLs by ", " instead of
by line-breaks.