Detect moves in (single) expanded file by default
On a diff page, if 'm' is pressed and a single file is expanded, let that
file be pre-selected in the drop-downs in the "Detect Moved Code" dialog
instead of the "Any" choice.  If no file is expanded, or if more than one
file is expanded, the "Any" choice is selected by default as before.