Include Critic version in administrator crash/error reports
Introduces sendAdministratorErrorReport() which sends an error
email to the adminstrators that also includes the currently
installed SHA1 of Critic itself. Used for exceptions raised
when generating pages or when processing operations plus when
a background service logs an ERROR.

sendAdministratorMessage() remains and doesn't include such
information, it is used for non-error report notifications
where the currently installed Critic SHA1 is not interesting
as context (e.g. when load limits are reached etc).