Tweak /endsession and /changepassword operations slightly
endsession: no need to check if the user is anonymous; the operation
doesn't set the "accept_anonymous_user" flag, so won't be called at all.
Also, it was buggy; if it had been called with an anonymous user, it
would have returned None, which would have caused a crash.

changepassword: support specifying the user whose password to manipulate
using either a user id or a user name.  Also rename the argument to
"subject" (since "user_id" makes little sense for a user name) and make
it optional (defaulting to the signed in user.)