Don't call /getrepositorypaths with "repository_name":null
For a Critic installation with 2 repos and a user with no filters,
the "Add Filter" dialog shows a Repository: SELECT with the 2 repos
plus an item called "Select a repository". If the user didn't pick
a repo but started to type in the Path: field directly, it would
fire off a /getrepositorypaths call with "repository_name" : null
causing an "Failed to count matched files" error to appear. This
fix makes sure that no calls to /getrepositorypaths are done before
a repository has been selected.