Drop "Indexes" and "MultiViews" options for /static-resource/
There's not much point in producing a directory listing of the
$INSTALL_DIR/resources/ directory, and we don't do content
negotiation, so enabling the "MultiViews" option doesn't make
much sense.

We do need the FollowSymLinks option though, since the jQuery
resources (such as jquery.js) are symlinks.

Also add test that checks that directory listing is disabled,
and that loading static resources works in general.