Improve rendering of commits and rebases on review front-page
If more than one rebase was listed at the top (not followed in time by
additional commits) then they were listed in reversed order.  This meant
the "[revert rebase]" link didn't work, since it tried to revert a rebase
that wasn't the most recent one.

Also fix the displayed author of the "commit" containing the changes
introduced in a rebase to be the user that performed the rebase.  It used
to display the author of the commit that was at the head of the review
branch after the rebase.

Also fix a rendering glitch when a review branch has had its history
rewritten in a way that simply reset the branch back to an earlier commit.
In that case, that earlier commit could be displayed at the wrong place,
and commits preceding it could be missing entirely form the commit list.