Make handling of fixup!/squash! in branch logs less strict
Always rewrite "fixup! FOO\n\nBAR" into "[fixup] BAR", even if FOO is not
found to reference any commit among the commits being output.  This makes
sense for instance if a review contains only (or starts with) a "fixup!"
commit of an earlier commit not in the review.  And this is just a bit of
UI sugar anyway; there's no particular reason to be strict and not apply
it for commits that are clearly fixups.

The one down-side of this is that commits whose summary is on the form

  fixup! FOO
  BAR

where FOO is not actually a reference to another commit, but rather a
description of this commit (with BAR being a continuation of that
description.)  This is of course an incorrect use of "fixup!", but it's
been known to happen.  Such commits will now be rendered (in a branch
log) as "[fixup] BAR" which will probably not make sense.