Produce nicer review summary when creating review via push
When a review is created via push, the summary of the single commit
is used as the (initial) summary of the review.  If the single commit
is a fixup or squash commit, it's more meaningful to use the second
(non-empty) line of the commit message as the review summary instead
of the fixup!/squash! line (which doesn't typically describe the
changes in the commit at all.)