Improve hint about fixing repository/database unsync
When a branch in the Git repository is not where it ought to be
according to the database, we reject pushes to it unless the push
resets the Git repository back to where the database thinks it
ought to be.  When rejecting, we output a hint about how to
reset the Git repository using "git push".  The hint wasn't to do
"git push -f", meaning it would sometimes fail.