Adjust behavior of testing framework context manager objects
Specifically the objects returned by the functions frontend.signin() and
repository.workcopy().  Instead of those functions doing the initial work
(signing in / creating the work copy) immediately, delay that to the
returned object's __enter__().  This mostly just Makes Sense<tm>.

(The context manager mechanism is designed to ensure that __exit__() runs
if __enter__() ran, so it makes sense that __exit__() then undoes what
__enter__() did, and thus that __enter__() actually does something.)