Introduce RestrictedString type for HTTP request params
To make critic crash proof when it comes to bad input at the
HTTP level, we're going to need quite a few checks for string
length and "allowed characters" whitelists. Instead of coding
this as lots of if statements in each operation, this new class
will allow us to do that using the existing type checker.