Fix forced update of tracked branch
In 820e5d14 ("Add support for githook customization") the function
deleteBranch() in index.py got an extra positional parameter.  The
call to it from updateBranch() in the case of a forced (non-ff)
update of a tracked branch was not adjusted to pass an extra
argument, so that call then crashed instead.