Fix crash when recreating tracked branch
When a tracked branch (with the 'force' flag set) is updated in a
non-ff fashion, it is internally first deleted and then recreated
(as if it was first deleted via 'git push <remote> :<name>' and
then pushed normally.)  The recreation part was broken, calling the
function createBranch() with user=None.

Fix by calling createBranches() instead, with a user name argument.
The function createBranches() is a thin wrapper (in the case of
creating a single branch) which handles the user name => user
conversion.