Add test for non-ff updates of tracked branches
Tests that a non-forced tracked branch stops tracking if the
source branch is rebased, and that a forced tracked branch
doesn't.

Also extend the /addtrackedbranch operation to take a 'forced'
parameter, and to return the added tracked branch's id.

Also make sure receive.deny{Deletes,NonFastforwards} are not
enabled in the main repository clone used for testing.  They
are not in Git's normal default settings, but this might be
overridden on the system running the tests.