Suppress "user-friendly" hook output for pushes by the system
This was already done when creating a branch.  Deleting a branch however
still output the "Deleted branch containing N commits." line.  This was
particularly unfortunate when the push was actually a non-ff update of a
forced tracked branch, since it was the only output, but the branch was
only temporary (and internally) deleted.

Instead, fix the output of more meaningful (for the logs) output in the
specific case of non-ff updates of tracked branches.  This was returned
from the updateBranch() method, but its return value was ignored by the
caller.  The correct way to produce output here is by writing it to the
stdout stream (sys.stdout is "redirected" and picked up by the caller.)