Set GIT_{AUTHOR,COMMITTER}_{NAME,EMAIL} variables more consistently
They need to be set when running Git commands that create commits (such
as 'git commit', 'git cherry-pick' and 'git merge') to get the commit
metadata correct.

For this purpose, add a utility function gitutils.getGitEnvironment().

This also fixes a crash when analyzing a non-ff review rebase that had
conflicts.