Fix crash when using [partial] filtering on rebased review
When rendering the "Squashed History" log, rebase information can
be used only if all commits in the review are included.  This is
always the case for the [reviewable], [relevant] and [full]
filters.  In the code, [full] and [partial] are both called
"files", with the difference being that [partial] also includes
URI parameters selecting a range of commits.  When deciding whether
to include rebase information when rendering the "Squashed History"
log, however, "files" was (incorrectly) assumed to be [full],
which caused a crash (assertion failure) if the filter was actually
[partial].