Extend 020-reviewrebase.py to check /showcommit's log as well
When /showcommit is loaded wia one of the filters [reviewable], [relevant]
and [full], the "Squashed History" log will be identical to the one on the
review's front-page, including information about rebases, so load it too
and apply the same check as to the on the review front-page's log.

This acts as a regression test for the "Fix handling of rebases in
"Squashed history" log display" commit in r/207.  (Not a perfect one; the
log is sometimes (randomly) rendered correctly even without the fix.)