Make unqualified searches always search in description and summary
This makes the different behaviours differ less from each other. What
happens is basically that you always search summary and description, but
if the term is a single word, it will also look in branches, and if it
looks like a path, it will also look at paths.

This change is also necessary for a coming simplification of the search UI.