Redesign the search page
The design is based on some assumptions:
 - Free text search is common and should search all relevant fields.
 - Owners is the most common user search, so that is the default.
 - If you want to search in a branch, you will often select a
   repository. Their relationship is made clear in this design, with a
   "this-then-that" workflow.

Slightly more Python code at the moment as there are no helpers for
non-table paleyellow sections, but in the long run it's a simpler approach.

The CSS uses flexbox, currently with no vendor prefixes. It is tested in
Opera 12 and Chromium-based browsers, but might be broken in Firefox,
Safari and IE10, but those should be quick fixes.