Sort migrations according to initial commit date before running
When upgrading over really large spans of commits one might be executing
a large set of migration scripts and in this case the execution order of
these are critically important.  This fix makes sure the migration scripts
execute in the order that they were originally added to the Critic git
repo.

An example of a bug caused by wrong migration script execution order can
be seen (along with other issues) in the upgrade from 30c4f10 to 25071da.
The two migration scripts that have an order dependency in this case is:

preference.commit.diff.rulerColumn.py (added 2013-04-08) which adds a
new pref and lists default_integer in its INSERT query, and then
dbschema.per-repository-or-filter-preferences.py (added 2013-07-10)
which deletes the default_integer column in preferences.