Fix out-of-repo symlink hashing
Critic currently uses "git hash-object" when finding the content hash of a
installed file (i.e. /usr/share/critic/whatever.py etc), and on the git repo
side it uses "git ls-tree".  This works well if the target is a normal file,
but it doesn't work if the installed target is a symlink.  This is because
git hash-object will follow the link and hash the content of the file that
the link points to, whereas "git ls-tree" returns the content hash of the
symlink itself.

"git hash-object" currently has no switch that instructs it to hash the link
itself when given a symlink.  Because of this we need to manually compute the
content hash of the symlink, in such a way that it matches the hash returned
by "git ls-tree".

Before this fix, when Critic was running an upgrade where a file pointed to
by a symlink changed (e.g. jquery.js), it would incorrectly say that "the
file has changed since it was installed" and ask the user for confirmation.

Such an issue can be seen in the upgrade from 30c4f10 to 25071da, after
various other issues is worked around.