Show "Create Review" button after reprocessing branch with new base
When a user pushes a branch with e.g. two commits to Critic, and then
amends to the second commit and pushes the result as a new second
branch; then the /branches page shows the second branch containing
just one commit but it also offers to reprocess Critic's view of that
branch by selecting a new "base" for it.  When this is done, such that
both commits are visible on the second branch, then the usual
"Create Review" button wasn't immediately visible post-reprocessing.
This fix makes sure that the page is reloaded so that a clean normal
view of the reprocessed branch is seen.