Adjust password UI on /home
If the user has no password, or if an administrator is manipulating
another user, set the button's label to "Set password" instead of
"Change password", and open a dialog with no "Current password" field.
(The latter was already the case if the user was an administrator.)

This fixes two things:

 - Confusing UI when a user tried to create a password: the dialog
   asked for the current password even if the user didn't have one.
   (And in the end it was also impossible to successfully create a
   password since the current password check wasn't disabled.)

 - There was no current password check for administrators changing
   their own password.  There's no reason for that; administrators
   should only be able to change other user's passwords without
   knowing them.

Also add an "Edit" button when an administrator user loads /home as
another user without readonly=no.