Set "display: block" on textareas in dialogs
This eliminates the need to make the textarea 3px "too low" to avoid
having scrollbars in the dialogs in Chrome.  Also set "outline: 0" on
the textareas to avoid Chrome's focus outline, which looks odd in these
textareas, and are also quite redundant.

Also some other minor changes to the calculation of available height
in dialogs and the styling of content in them that doesn't change the
end result (much.)