Fix retrieval of deleted file content for merge commits
The fix makes sure that the right blob hash (corresponding
the the currently selected parent) is passed to the
"fetchlines" operation when the "Fetch Deleted Lines"
button is clicked. Before this fix that button would
just hit a JS TypeError.

For example, if FILENAME existed with different content in
two different commits and a merge of those two commits
resolves the conflict by simple deleting the file, then the
"Fetch Deleted Lines" button should show either versions of
the old file content depending on which parent is selected.

A variant of this issue happens if A) a review branch has added
a file called NEWFILE and has changed a file called CONFLICTING,
and B) this review branch is rebased to a new upstream which
has different changes in CONFLICTING causing a conflict during
the rebase, and C) for whatever reason the person who is doing
the rebase decides to also delete NEWFILE. If that is done then
then fake merge commit that Critic generates to represent the
new-upstream-rebase will exhibit the same problem.