Add support for "pause, test, upgrade" loop in testing framework
When the --pause-upgrade-loop argument is used in conjunction with
one of the other --pause* arguments that makes testing pause, instead
of continuing the testing on ENTER, push HEAD in the local repository
to the tested system and upgrade it to that commit, staying paused.

A command to run locally before pushing HEAD can also be specified.
If this is something like "git commit -a -mWIP", this mechanism can be
rather convenient to use to quickly make changes and test them in an
installed system.