During installation, add administrator user using criticctl
This is quite a bit simpler.  And since we depended on the (installed)
'auth' module already, we might as well depend on more installed things.
We just need to make sure to install the criticctl executable a bit
earlier, which is uncomplicated since the installation of it doesn't
depend on anything but knowing installation paths.