Fix criticctl's global (non-command) argument handling
The division into "global arguments", "command" and "command arguments"
in the criticctl script was overly simple; treating the first argument
not starting with a dash as the command, everything before as global
arguments and everything after as command arguments.  This meant that
with "criticctl -i main foo", "main" was interpreted as the command, and
"-i" as the global arguments, which would fail, since "-i" requires a
value.

Also add a test for the basic criticctl functionality, and when doing so,
renumber the existing criticctl tests +1.