Restructure "Git emails" handling to allow multiple Critic users
IOW, allow multiple Critic users to claim to be using a particular email
address in Git, and thus be considered authors of the same commits.  The
only situation in which this ought to happen is if a person has multiple
Critic users in the same system.

Also simplify some code by using the CommitUserTime.getFullname() helper
function.  This helper now returns the Critic user's fullname if a single
Critic user is identified by the Git email address; otherwise it returns
the name from the Git meta-data.