Signal test failure if test logs errors
In some cases, it is meaningful to log an error and continue testing;
typically because the error wouldn't affect the rest of the test.

If a test did this, it would not be considered a failure.  This was
mostly not a problem, but broke --pause-on-failure, and caused an
incorrect "tests failed" counter at the end of the test run.

CriticTester only looks at logged errors (and also warnings) when
determining whether a test run is successful, so it was unaffected by
this glitch.