Fix crash when displaying review front-page of some rebased reviews
If a review was first rebased (history rewrite) in such a way that the
review branch was simply reset to a previous point, and then immediately
following that (with no additional commits pushed in between) was rebased
(non-ff move) again, rendering the review front-page would crash.

Technically, we failed to recognize that both rebases were "top" rebases
that should be displayed at the top of the log, before the first real
commit, because the old head of the first rebase (the move) was a commit
that was part of the review.  We assumed that this meant this commit
should be displayed immediately after that rebase, which was not the case.

This log rendering code must die...