Fix diff on /showcomment in some rare situations
Even if reviewing.html.renderCodeCommentChain() decided that the comment
should be on the left ("old") side of the diff, diff.context.ContextLines
could decide it belonged on the right ("new") side of the diff instead,
if the chain was connected to a set of lines in the "new" file version.
Avoid this by explicitly telling diff.context.ContextLines whether the
comments belong on the left side.  (Showing comments on the right side,
when possible, is the default.)

IIRC, the asymmetry between the two parts of the code could lead to
diff.context.ContextLines not including the lines on the left, where the
comment markers should be, in the diff, if the comment was connected to
lines on the right side that were sufficiently far away.