Fully expose TrackedBranch API
Was previously available in a limited read-only form as the
'trackedBranch' property on Review objects, for reviews tracking a remote
branch.

The TrackedBranch class is now exported on the critic module, and now has
the static TrackedBranch.find() function for finding arbitrary tracked
branches.

Also added a TrackedBranch.triggerUpdate() method for scheduling an
immediate update of the branch, TrackedBranch.{enable,disable} methods
for enabling and disabling the tracking, and TrackedBranch.getLogEntry()
for fetching the log entry (and pre-receive hook output) for the update
to a particular commit.