Use operation.User for subject argument to some operations
The operations are /setgitemails, /changepassword and /addemailaddress,
all of which we have tests for.

The argument to /setgitemails was renamed from "user_id" to "subject" (or
"subject_id"/"subject_name".)

Also delete the VerifyEmailAddress class, which was never finished and not
hooked up.  (The use-case was that an administrator would be able to use
it to force-verify another user's unverified email address, but this never
happened.)