Move some imports around to depend less on having an installed system
It's nice, for unit testing purposes, if some basic modules can be
imported even if not running in an installed system, despite them not
really including that much functionality that is usable in that setting.

This commit makes it so, and adds unit tests that simply check that the
modules can be imported.