Make git-send-pack output suffix stripping more reliable
We deleted all occurrences of "\x1b[K" in the output from "git push", but
this is not the suffix actually used under all circumstances; if $TERM is
"dumb", the suffix is 8 spaces instead.

We now set $TERM to "dumb", and strip 8 spaces from the end of lines, if
there are 8 spaces there.