Fix leaking "git cat-file --batch" processes
When an extension process shuts down "uncleanly" (i.e. without killing its
"git cat-file --batch" processes first,) those Git processes where left
alive indefinitely.  The reason for that was that a reference to the
writing end of the pipe used as Git's stdin was inherited into the child
process and thus never fully closed.

Avoid this by setting the FD_CLOEXEC flag on the file descriptor.

Also update the v8-jsshell submodule pointer to the commit that adds the
IO.File.setCloseOnExec() method that we use to do that.