Use the correct SHA-1 when constructing v8-jsshell cache file path
The current sub-module SHA-1 checked out in the repository in which the
tests were run was used, but this is wrong if --commit is used to override
the tested commit, and also wrong if extend() runs before upgrade().

Keep track of which commit is current installed as 'self.current_commit',
and use that when constructing the cache file path.