Disable v8-jsshell caching if upgrading the v8-jsshell sub-module
If the v8-jsshell sub-module is modified in an upgrade, upgrade.py runs
extend.py, and "extend.py --fetch" fails if we used a cached v8-jsshell,
since the sub-module directory isn't empty then (we copy the cached
executable into it.)

It's meaningful to test the (automatic) incremental recompilation of
v8-jsshell that happens as part of the upgrade, so simply disable the
caching in this case.

This means that the testing of such commits becomes quite expensive, but
this should be sufficiently offset by the fact that sub-module updates
are rare.